+++ MegaDeskWeb-Angeles/Pages/Quotes/Edit.cshtml.cs
@@ -23,9 +23,9 @@ namespace MegaDeskWeb_Angeles.Pages.Quotes
         [BindProperty]
         public DeskQuote DeskQuote { get; set; }
 
-        public int calcDrawerCost() => DeskQuote.Drawers * 50;
+        public decimal calcDrawerCost() => DeskQuote.Drawers * 50;
 
-        public int calcMaterialCost()
+        public decimal calcMaterialCost()
         {
             switch (DeskQuote.Material)
 
@@ -57,7 +57,7 @@ namespace MegaDeskWeb_Angeles.Pages.Quotes
 
             }
         }
-        public int calcSurfaceAreaCost()
+        public decimal calcSurfaceAreaCost()
         {
 
             if (DeskQuote.SurfaceArea > 1000)
@@ -70,7 +70,7 @@ namespace MegaDeskWeb_Angeles.Pages.Quotes
                 return 0;
             }
         }
-        public int calcRushOrderCost()
+        public decimal calcRushOrderCost()
 
         {
 
@@ -183,7 +183,7 @@ namespace MegaDeskWeb_Angeles.Pages.Quotes
             }
 
         }
-        public int calcQuoteTotal()
+        public decimal calcQuoteTotal()
         {
             return DeskQuote.DrawerCost + DeskQuote.MaterialCost + DeskQuote.SurfaceAreaCost + DeskQuote.RushCost;
         }
